Move some code in checkpointFile for more safety
There is no reason for Writer to have access to Reader functions and
visa versa
